+++ HISTORY.rst
@@ -157,6 +157,12 @@
 - Remove unused `ModelFormFieldFilteredModelChoice`.
 - Make `FormMixinFieldsets` a subclass of `FormMixinFormMeta`.
 
+4.9.6 (XXXX-XX-XX)
+==================
+- Rename file metadata model mixin class for clarity.
+- Consolidate dynamic filtering field reload code. Add the new
+  `FormMixinFilteredFieldsReload` mixin.
+
 4.9.5 (2025-09-28)
 ==================
 - Improvements and changes from version 4.8.8.